+++ src/has.js
@@ -1,3 +1,5 @@
+import { both, either, eq, test } from 'kyanite'
+
 /**
  * @name hasValue
  * @since v1.0.0
@@ -8,10 +10,10 @@
  * @returns {Boolean} Returns true or false based on the validation test
  *
  * @example
- * const result = hasValue('11'); // => true
- * const result = hasValue(''); // => false
+ * const result = hasValue('11') // => true
+ * const result = hasValue('') // => false
  */
-export const hasValue = val => val === 0 || Boolean(val)
+export const hasValue = either(eq(0), Boolean)
 
 /**
  * @name hasNumbers
@@ -23,11 +25,11 @@ export const hasValue = val => val === 0 || Boolean(val)
  * @returns {Boolean} Returns true or false based on the validation test
  *
  * @example
- * const result = hasNumbers('11'); // => true
- * const result = hasNumbers('eew2211'); // => true
- * const result = hasNumbers('eerrt'); // => false
+ * const result = hasNumbers('11') // => true
+ * const result = hasNumbers('eew2211') // => true
+ * const result = hasNumbers('eerrt') // => false
  */
-export const hasNumbers = val => (/[0-9]/).test(val)
+export const hasNumbers = test(/[0-9]/)
 
 /**
  * @name hasLetters
@@ -39,11 +41,11 @@ export const hasNumbers = val => (/[0-9]/).test(val)
  * @returns {Boolean} Returns true or false based on the validation test
  *
  * @example
- * const result = hasLetters('11'); // => false
- * const result = hasLetters('eew2211'); // => true
- * const result = hasLetters('eerrt'); // => true
+ * const result = hasLetters('11') // => false
+ * const result = hasLetters('eew2211') // => true
+ * const result = hasLetters('eerrt') // => true
  */
-export const hasLetters = val => (/[A-Z]/i).test(val)
+export const hasLetters = test(/[A-Z]/i)
 
 /**
  * @name hasSpecialCharacters
@@ -55,11 +57,11 @@ export const hasLetters = val => (/[A-Z]/i).test(val)
  * @returns {Boolean} Returns true or false based on the validation test
  *
  * @example
- * const result = hasSpecialCharacters('11%%$#'); // => true
- * const result = hasSpecialCharacters('eew2211!@'); // => true
- * const result = hasSpecialCharacters('eerrt'); // => false
+ * const result = hasSpecialCharacters('11%%$#') // => true
+ * const result = hasSpecialCharacters('eew2211!@') // => true
+ * const result = hasSpecialCharacters('eerrt') // => false
  */
-export const hasSpecialCharacters = val => (/\W/).test(val)
+export const hasSpecialCharacters = test(/\W/)
 
 /**
  * @name hasNumbersOrSpecials
@@ -71,11 +73,11 @@ export const hasSpecialCharacters = val => (/\W/).test(val)
  * @returns {Boolean} Returns true or false based on the validation test
  *
  * @example
- * const result = hasNumbersOrSpecials('11%%$#'); // => true
- * const result = hasNumbersOrSpecials('eew2211!@'); // => true
- * const result = hasNumbersOrSpecials('eerrt'); // => false
+ * const result = hasNumbersOrSpecials('11%%$#') // => true
+ * const result = hasNumbersOrSpecials('eew2211!@') // => true
+ * const result = hasNumbersOrSpecials('eerrt') // => false
  */
-export const hasNumbersOrSpecials = val => hasNumbers(val) || hasSpecialCharacters(val)
+export const hasNumbersOrSpecials = either(hasNumbers, hasSpecialCharacters)
 
 /**
  * @name hasUpperAndLowerCase
@@ -87,8 +89,8 @@ export const hasNumbersOrSpecials = val => hasNumbers(val) || hasSpecialCharacte
  * @returns {Boolean} Returns true or false based on the validation test
  *
  * @example
- * const result = hasUpperAndLowerCase('11%%$#'); // => false
- * const result = hasUpperAndLowerCase('Eew2211!@'); // => true
- * const result = hasUpperAndLowerCase('eERrt'); // => true
+ * const result = hasUpperAndLowerCase('11%%$#') // => false
+ * const result = hasUpperAndLowerCase('Eew2211!@') // => true
+ * const result = hasUpperAndLowerCase('eERrt') // => true
  */
-export const hasUpperAndLowerCase = val => (/[A-Z]/).test(val) && (/[a-z]/).test(val)
+export const hasUpperAndLowerCase = both(test(/[A-Z]/), test(/[a-z]/))