+++ src/is.js
@@ -1,85 +1,6 @@
-import { between, curry } from 'kyanite'
-
-/**
- * @name isLongerThan
- * @since v4.0.0
- * @function
- * @category Is
- * @description Validates if the value is longer or equal to the minLen option
- * @param {Number} rule The length the value needs to be greater than or equal to
- * @param {String} val The value to validate against
- * @returns {Boolean} Returns true or false based on the validation test
- *
- * @example
- * isLongerThan(5, 'the cow ate grass') // => true
- * isLongerThan(5, 'the') // => false
- *
- * // OR
- *
- * const check = isLongerThan(5)
- *
- * check('the cow ate grass') // => true
- * check('the') // => false
- */
-export const isLongerThan = curry((minLen, val) => val.length >= minLen)
-
-/**
- * @name isShorterThan
- * @since v4.0.0
- * @function
- * @category Is
- * @description Validates if the value is shorter or equal to the maxLen option
- * @param {Number} rule The length the value needs to be Less than or equal to
- * @param {String} val The value to validate against
- * @returns {Boolean} Returns true or false based on the validation test
- *
- * @example
- * isShorterThan(5, 'the') // => true
- * isShorterThan(5, 'the cow ate grass') // => false
- *
- * // OR
- *
- * const check = isShorterThan(5)
- *
- * check('the') // => true
- * check('the cow ate grass') // => false
- */
-export const isShorterThan = curry((maxLen, val) => val.length <= maxLen)
-
-/**
- * @name isCorrectLength
- * @since v4.0.0
- * @function
- * @category Is
- * @description Validates if the value is between or equal to the maxLen and minLen options
- * @param {Object} rule Object containing the two numbers to stay between
- * @param {Number} rule.maxLen The length the value needs to be Less than or equal to
- * @param {Number} rule.minLen The length the value needs to be greater than or equal to
- * @param {String} val The value to validate against
- * @returns {Boolean} Returns true or false based on the validation test
- *
- * @example
- * isCorrectLength({
- *  maxLen: 10,
- *  minLen: 1
- * }, 'the') // => true
- * isCorrectLength({
- *  maxLen: 10,
- *  minLen: 1
- * }, 'the cow ate grass') // => false
- *
- * // OR
- *
- * const check = isCorrectLength({
- *  maxLen: 10,
- *  minLen: 1
- * })
- *
- * check('the') // => true
- * check('the cow ate grass') // => false
- */
-export const isCorrectLength = curry(({ maxLen, minLen }, val) =>
-  isLongerThan(minLen, val) && isLongerThan(maxLen, val))
+import { between, test } from 'kyanite'
+import _curry2 from './internal/_curry2'
+import _curry3 from './internal/_curry3'
 
 /**
  * @name isDate
@@ -93,8 +14,7 @@ export const isCorrectLength = curry(({ maxLen, minLen }, val) =>
  * @example
  * isDate('1/2/2019') // => true
  */
-export const isDate = val =>
-  (/^((1[0-2])|(0?[1-9]))[-/.]?((0?[1-9])|([1-2][0-9])|(3[0-1]))[-/.]?(([1-2]{1}[0-9]{3})|([0-9]{2}))$/m).test(val)
+export const isDate = test(/^((1[0-2])|(0?[1-9]))[-/.]?((0?[1-9])|([1-2][0-9])|(3[0-1]))[-/.]?(([1-2]{1}[0-9]{3})|([0-9]{2}))$/m)
 
 /**
  * @name isDateShort
@@ -109,7 +29,7 @@ export const isDate = val =>
  * isDateShort('1/19') // => true
  * isDateShort('13/19') // => false
  */
-export const isDateShort = val => (/^((1[0-2])|(0?[1-9]))[-/.]?((0?[1-9])|([1-2][0-9])|(3[0-1]))[-/.]?$/m).test(val)
+export const isDateShort = test(/^((1[0-2])|(0?[1-9]))[-/.]?((0?[1-9])|([1-2][0-9])|(3[0-1]))[-/.]?$/m)
 
 /**
  * @name isDateProper
@@ -123,8 +43,7 @@ export const isDateShort = val => (/^((1[0-2])|(0?[1-9]))[-/.]?((0?[1-9])|([1-2]
  * @example
  * isDateProper('2019/1/2') // => true
  */
-export const isDateProper = val =>
-  (/^(([1-2]{1}[0-9]{3})|([0-9]{2}))[-/.]?((1[0-2])|(0?[1-9]))[-/.]?((0?[1-9])|([1-2][0-9])|(3[0-1]))$/m).test(val)
+export const isDateProper = test(/^(([1-2]{1}[0-9]{3})|([0-9]{2}))[-/.]?((1[0-2])|(0?[1-9]))[-/.]?((0?[1-9])|([1-2][0-9])|(3[0-1]))$/m)
 
 /**
  * @name isEmail
@@ -138,11 +57,7 @@ export const isDateProper = val =>
  * @example
  * isEmail('dusty@gmail.com') // => true
  */
-export const isEmail = val => {
-  const emailRegex = /^[\w\u00c0-\u017f][\w.-_\u00c0-\u017f]*[\w\u00c0-\u017f]+[@][\w\u00c0-\u017f][\w.-_\u00c0-\u017f]*[\w\u00c0-\u017f]+\.[a-z]{2,4}$/i
-
-  return emailRegex.test(val)
-}
+export const isEmail = test(/^[\w\u00c0-\u017f][\w.-_\u00c0-\u017f]*[\w\u00c0-\u017f]+[@][\w\u00c0-\u017f][\w.-_\u00c0-\u017f]*[\w\u00c0-\u017f]+\.[a-z]{2,4}$/i)
 
 /**
  * @name isNumber
@@ -165,7 +80,7 @@ export const isNumber = val => !isNaN(val)
  * @function
  * @category Is
  * @description Validates if a provided value is a positive number
- * @param {String|Number} val The value to validate against
+ * @param {Number} val The value to validate against
  * @returns {Boolean} Returns true or false based on the validation test
  *
  * @example
@@ -173,7 +88,7 @@ export const isNumber = val => !isNaN(val)
  * isPositive(2) // => true
  * isPositive(-2) // => false
  */
-export const isPositive = val => !isNaN(val) && Number(val) >= 0
+export const isPositive = val => !isNaN(val) && val >= 0
 
 /**
  * @name isNegative
@@ -181,7 +96,7 @@ export const isPositive = val => !isNaN(val) && Number(val) >= 0
  * @function
  * @category Is
  * @description Validates if a provided value is a negative number
- * @param {String|Number} val The value to validate against
+ * @param {Number} val The value to validate against
  * @returns {Boolean} Returns true or false based on the validation test
  *
  * @example
@@ -189,7 +104,7 @@ export const isPositive = val => !isNaN(val) && Number(val) >= 0
  * isNegative(-2) // => true
  * isNegative(2) // => false
  */
-export const isNegative = val => !isNaN(val) && Number(val) < 0
+export const isNegative = val => !isNaN(val) && val < 0
 
 /**
  * @name isVin
@@ -203,11 +118,7 @@ export const isNegative = val => !isNaN(val) && Number(val) < 0
  * @example
  * vin('JM1CW2BL8C0127808') // => true
  */
-export const isVin = val => {
-  const vinRegex = /^[a-hj-npr-z0-9]{9}[a-hj-npr-tv-y1-9]{1}[a-hj-npr-z0-9]{7}$/i
-
-  return vinRegex.test(val)
-}
+export const isVin = test(/^[a-hj-npr-z0-9]{9}[a-hj-npr-tv-y1-9]{1}[a-hj-npr-z0-9]{7}$/i)
 
 /**
  * @name isZip
@@ -222,7 +133,7 @@ export const isVin = val => {
  * isZip('44444') // => true
  * isZip('232') // => false
  */
-export const isZip = val => (/^\d{5}(-\d{4})?$/).test(val)
+export const isZip = test(/^\d{5}(-\d{4})?$/)
 
 /**
  * @name isCAPostalCode
@@ -237,7 +148,7 @@ export const isZip = val => (/^\d{5}(-\d{4})?$/).test(val)
  * isCAPostalCode('K1A0B1') // => true
  * isCAPostalCode('44444') // => false
  */
-export const isCAPostalCode = val => (/^[ABCEGHJKLMNPRSTVXY]{1}\d{1}[A-Z]{1} *\d{1}[A-Z]{1}\d{1}$/i).test(val)
+export const isCAPostalCode = test(/^[ABCEGHJKLMNPRSTVXY]{1}\d{1}[A-Z]{1} *\d{1}[A-Z]{1}\d{1}$/i)
 
 /**
  * @name isPhone
@@ -252,7 +163,7 @@ export const isCAPostalCode = val => (/^[ABCEGHJKLMNPRSTVXY]{1}\d{1}[A-Z]{1} *\d
  * isPhone('555-666-7777') // => true
  * isPhone('5556667777') // => true
  */
-export const isPhone = (val = '') => (/^[0-9]{10}$/).test(val.replace(/\W/g, ''))
+export const isPhone = (val = '') => test(/^[0-9]{10}$/, val.replace(/\W/g, ''))
 
 /**
  * @name isLicensePlate
@@ -268,7 +179,7 @@ export const isPhone = (val = '') => (/^[0-9]{10}$/).test(val.replace(/\W/g, '')
  * isLicensePlate('SSS-1829') // => true
  * isLicensePlate('SSSS 188') // => false
  */
-export const isLicensePlate = val => (/^([A-Z]|[0-9]){1,3}(\s|-|•)?([A-Z]|[0-9]){3,5}$/i).test(val)
+export const isLicensePlate = test(/^([A-Z]|[0-9]){1,3}(\s|-|•)?([A-Z]|[0-9]){3,5}$/i)
 
 /**
  * @name isBelowMax
@@ -277,25 +188,19 @@ export const isLicensePlate = val => (/^([A-Z]|[0-9]){1,3}(\s|-|•)?([A-Z]|[0-9
  * @category Is
  * @description Validates if a provided value is a below the set maximum
  * @param {Number} m The max to validate against
- * @param {String} val The value to validate with
+ * @param {Number} val The value to validate with
  * @returns {Boolean} Returns true or false based on the validation test
  *
  * @example
  * const below = isBelowMax(20)
- * below('19') // => true
+ * below(19) // => true
  *
  * // OR
  *
- * isBelowMax(20, '19') // => true
- * isBelowMax(20)('19') // => true
+ * isBelowMax(20, 19) // => true
+ * isBelowMax(20)(19) // => true
  */
-export const isBelowMax = curry((m, val) => {
-  if (m.max != null) {
-    return !isNaN(val) && Number(val) < m.max
-  }
-
-  return !isNaN(val) && Number(val) < m
-})
+export const isBelowMax = _curry2((m, val) => !isNaN(val) && val < m)
 
 /**
  * @name isAboveMin
@@ -304,25 +209,19 @@ export const isBelowMax = curry((m, val) => {
  * @category Is
  * @description Validates if a provided value is a below the set minimum
  * @param {Number} m The min to validate against
- * @param {String} val The value to validate with
+ * @param {Number} val The value to validate with
  * @returns {Boolean} Returns true or false based on the validation test
  *
  * @example
  * const above = isAboveMin(15)
- * above('19') // => true
+ * above(19) // => true
  *
  * // OR
  *
- * isAboveMin(15, '19') // => true
- * isAboveMin(15)('19') // => true
+ * isAboveMin(15, 19) // => true
+ * isAboveMin(15)(19) // => true
  */
-export const isAboveMin = curry((m, val) => {
-  if (m.min != null) {
-    return !isNaN(val) && Number(val) > m.min
-  }
-
-  return !isNaN(val) && Number(val) > m
-})
+export const isAboveMin = _curry2((m, val) => !isNaN(val) && val > m)
 
 /**
  * @name isBetween
@@ -330,21 +229,18 @@ export const isAboveMin = curry((m, val) => {
  * @function
  * @category Is
  * @description Checks if the provided value is between the max and min
- * @param {Object} m The object containing max and min values
- * @param {Number} m.max The maximum the value should be below
- * @param {Number} m.min The minimum the value should be above
+ * @param {Number} min The minimum the value should be above
+ * @param {Number} max The maximum the value should be below
  * @param {Number} val The value to compare
  * @returns {Boolean} Whether or not the number is between the max and min numbers
  * @example
- * isBetween({ max: 10, min: 5 }, 6) // => true
- * isBetween({ max: 10, min: 5 }, 3) // => false
+ * isBetween(5, 10, 6) // => true
+ * isBetween(5, 10, 3) // => false
  *
  * // It's also curried
- * const fn = isBetween({ max: 10, min: 5 })
+ * const fn = isBetween(5, 10)
  *
  * fn(6) // => true
  * fn(3) // => false
  */
-const _isBetween = (m, val) => !isNaN(val) && between(m.min, m.max, val)
-
-export const isBetween = curry(_isBetween)
+export const isBetween = _curry3((min, max, val) => !isNaN(val) && between(min, max, val))