+++ src/no.js
@@ -1,3 +1,5 @@
+import { compose, isNil, match } from 'kyanite'
+
 /**
  * @name noSpecials
  * @since v1.0.0
@@ -8,11 +10,11 @@
  * @returns {Boolean} Returns true or false based on the validation test
  *
  * @example
- * noSpecials('AAAA'); // => true
+ * noSpecials('AAAA') // => true
  * noSpecials('1122334') // => true
- * noSpecials('AAA12!#$'); // => false
+ * noSpecials('AAA12!#$') // => false
  */
-export const noSpecials = (val = '') => String(val).match(/\W/) === null
+export const noSpecials = compose(isNil, match(/\W/))
 
 /**
  * @name noNumbers
@@ -24,11 +26,11 @@ export const noSpecials = (val = '') => String(val).match(/\W/) === null
  * @returns {Boolean} Returns true or false based on the validation test
  *
  * @example
- * noNumbers('AAAA'); // => true
+ * noNumbers('AAAA') // => true
  * noNumbers('1122334') // => false
- * noNumbers('AAA12!#$'); // => false
+ * noNumbers('AAA12!#$') // => false
  */
-export const noNumbers = (val = '') => String(val).match(/[0-9]/) === null
+export const noNumbers = compose(isNil, match(/[0-9]/))
 
 /**
  * @name noLetters
@@ -41,7 +43,7 @@ export const noNumbers = (val = '') => String(val).match(/[0-9]/) === null
  *
  * @example
  * noLetters('1122334') // => true
- * noLetters('AAAA'); // => false
- * noLetters('AAA12!#$'); // => false
+ * noLetters('AAAA') // => false
+ * noLetters('AAA12!#$') // => false
  */
-export const noLetters = (val = '') => String(val).match(/[A-Z]/i) === null
+export const noLetters = compose(isNil, match(/[A-Z]/i))